Add support for auto-dim-other-buffers.el
Auto-dim-other-buffers.el is a minor mode that provides a visual
distinction between active and inactive buffers by modifying face
attributes. The source code for auto-dim-other-buffers.el is available
at https://github.com/mina86/auto-dim-other-buffers.el.

In order to implement support for auto-dim-other-buffers.el in the
gruvbox theme, this commit introduces, in addition to the gruvbox-bg
colour that contains the background colour under the current theme, a
secondary gruvbox-bg_inactive colour that is always one level
softer. In this way, auto-dim-other-buffers.el with gruvbox provides a
subtle indication of active and inactive buffers, where inactive
buffers have minimally lower contrast.

In addition, this commit harmonizes the colour definitions among the
soft, hard and medium flavours so that every flavour provides the full
set of dark0_hard, dark0 and dark0_soft variant colours for theming.