Make some smart-mode-line info visible in light theme (#137)
The current color for the sml/global face (used for rendering e.g.
column indicator, minor mode names) was gruvbox-burlywood4. This color
doesn't invert and it's invisible in a focused window when using the
light hard variant.

This fixes this by using gruvbox-light4, which will invert correctly
and display properly in light themes.