feat: add corfu support
Corfu is a new package which instead of trying to define the completion
API (like company.el), uses the default completion-at-point-funcions
(a.k.a cafs), which are normally defined by programming modes.

Corfu also uses a child-frame, which is one of the star features of
Emacs 26.

I added corfu support to gruvbox, using the same principles used in the
company face setup.