feat: add ledger-mode support
Ledger-mode is a plain accounting mode created by the ledger cli
maintainer. It works with the double accounting method.

The problem with the default faces for ledger-mode was that the ledger
transactions get highlighted when the cursor is over them, hence
the ledger-font-xact-highlight-face. Before the change, the transactions
were highlighted with a black face, hiding all inside content.

We now default to 'gruvbox-dark1', which seems like a very sane choice.