feat: add marginalia support
Marginalia adds metadata from different types of targets to the
minibuffer, pairing well with other packages like Corfu, Consult
and Embark.

The inherit faces that Marginalia defines work already pretty well
out-of-the-box with Gruvbox, but the 'marginalia-documentation' is
pretty much hidden underneath the 'highlight' face. Making the
font brighter and italic is a great solution.