Add font-lock-number and font-lock-property faces
These two faces are used with font-lock to fontify different aspects of
the buffers. In particular, if using the new treesit feature in Emacs
29, numbers on tree-sitter enabled buffers will appear blank.

Also, considering members use bright_purple, let's stick
to bright_purple on properties too. It was defaulting to the built-in
one, making the fontified buffer really "orange" on dark mode.